move tree store manifest to tree module
