+++ src/packages/settings/data-types/repository/manifests.ts
@@ -1,11 +1,10 @@
 import { UmbDataTypeRepository } from './data-type.repository.js';
 import { UmbDataTypeStore } from './data-type.store.js';
-import { UmbDataTypeTreeStore } from '../tree/data-type.tree.store.js';
 import { manifests as itemManifests } from './item/manifests.js';
 import { manifests as moveManifests } from './move/manifests.js';
 import { manifests as copyManifests } from './copy/manifests.js';
 import { manifests as folderManifests } from './folder/manifests.js';
-import type { ManifestStore, ManifestTreeStore, ManifestRepository } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestStore, ManifestRepository } from '@umbraco-cms/backoffice/extension-registry';
 
 export const DATA_TYPE_REPOSITORY_ALIAS = 'Umb.Repository.DataType';
 
@@ -17,7 +16,6 @@ const repository: ManifestRepository = {
 };
 
 export const DATA_TYPE_STORE_ALIAS = 'Umb.Store.DataType';
-export const DATA_TYPE_TREE_STORE_ALIAS = 'Umb.Store.DataTypeTree';
 
 const store: ManifestStore = {
 	type: 'store',
@@ -26,19 +24,4 @@ const store: ManifestStore = {
 	api: UmbDataTypeStore,
 };
 
-const treeStore: ManifestTreeStore = {
-	type: 'treeStore',
-	alias: DATA_TYPE_TREE_STORE_ALIAS,
-	name: 'Data Type Tree Store',
-	api: UmbDataTypeTreeStore,
-};
-
-export const manifests = [
-	repository,
-	store,
-	treeStore,
-	...itemManifests,
-	...moveManifests,
-	...copyManifests,
-	...folderManifests,
-];
+export const manifests = [repository, store, ...itemManifests, ...moveManifests, ...copyManifests, ...folderManifests];
